pyproject.toml
[project]
name = "personal-chef"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "DapaMassimo", email = "massimo.dapa96@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "langchain==1.2.8",
    "langchain-openai>=1.1.7",
    "langchain-tavily>=0.2.17",
    "python-dotenv>=1.2.1",
    "tavily>=1.1.0",
    "tavily-python>=0.7.21",
]

[build-system]
requires = ["uv_build>=0.9.5,<0.10.0"]
build-backend = "uv_build"
